Changed PE to use floats to store the pin values. So now the valid interval for both analog and digital pins is 0..1, which I believe is more consistent.

+++ c++/UziFirmware/PE.h
@@ -1,13 +1,15 @@
 #pragma once
 
+
+
 #include "Arduino.h"
 
 // The following macros will work for standard arduino, other versions should redefine.
-#define TOTAL_PINS								        18
-#define ARRAY_INDEX(pin)					               ((pin) - 2)
-#define PIN_NUMBER(index)                                                    ((index) + 2)
-#define IS_ANALOG(pin)					      ((pin) >= 14 && (pin) <= 19)
-#define IS_DIGITAL(pin)						         (!IS_ANALOG(pin))
+#define TOTAL_PINS														   18
+#define ARRAY_INDEX(pin)										  ((pin) - 2)
+#define PIN_NUMBER(index)										((index) + 2)
+#define IS_ANALOG(pin)							 ((pin) >= 14 && (pin) <= 19)
+#define IS_DIGITAL(pin)										(!IS_ANALOG(pin))
 
 class PE {
 
@@ -22,15 +24,16 @@ public:
 	
 	unsigned char getMode(unsigned int);
 	void setMode(unsigned int, unsigned char);
-	unsigned short getValue(unsigned int);
-	void setValue(unsigned int, unsigned short);
+	float getValue(unsigned int);
+	void setValue(unsigned int, float);
 	long getMillis(void);
 	void delayMs(unsigned long);
 
 private:
 	
-	unsigned short _pinValues[TOTAL_PINS];
+	float _pinValues[TOTAL_PINS];
 	unsigned char _pinModes[TOTAL_PINS];
 
 };
+
